回転軸を決定するカスタムフック
